+++ src/device/index.ts
@@ -1,0 +1,1 @@
+export { AirfiDevice } from './AirfiDevice';